[Store] Split promotion unit tests
