+++ mooncake-store/tests/CMakeLists.txt
@@ -67,7 +67,12 @@ add_store_test(master_service_tenant_quota_test
 add_store_test(batch_remove_test batch_remove_test.cpp)
 add_store_test(master_service_ssd_test master_service_ssd_test.cpp)
 add_store_test(offload_on_evict_test offload_on_evict_test.cpp)
-add_store_test(promotion_on_hit_test promotion_on_hit_test.cpp)
+add_store_test(promotion_on_hit_test
+               master_service/promotion/admission_test.cpp)
+add_store_test(promotion_task_lifecycle_test
+               master_service/promotion/task_lifecycle_test.cpp)
+add_store_test(promotion_cleanup_metrics_test
+               master_service/promotion/cleanup_metrics_test.cpp)
 add_store_test(file_storage_promotion_test file_storage_promotion_test.cpp)
 add_store_test(master_service_ssd_test_for_snapshot
                ha/snapshot/master_service_ssd_test_for_snapshot.cpp)
